Get order of reactions in TUV
New Project.toml and Manifest.toml with version set to 0.1.0-DEV, Julia
version set to ≥0.7, and filehandling added as dependency.
In TUVtools.jl, new function setrxns to eventually write TUV input files
with an auto-generated photolysis reactions part. New function
findrxnlabel to find subroutine names in the rxn files, and create a
dictionary with the reaction labels used in each subroutine. New
function getorder to get the order of subroutine calls used in TUV and,
hence, the order of reactions in the TUV input files.
New file src/data/MCM-GECKO-A.db with  links of MCM/GECKO-A/DSMACC
reaction numbers to corresponding TUV labels (needed for setting flags
later).