Auto-generate TUV input files
Add package Printf in Project.toml.
New and updated functions to write to autogenerate TUV input files for
the TUV rxn files and the files with the subroutine calls.
Choose options to leave flags untouched (negative setflags), set all
flags to F (setflags = 0), T (setflags = 1) or set the reactions used in
MCM/GECKO-A to T (setflags = 2).