Generate wiki files with reaction numbers
New exported function generate_wiki in TUVtools.jl to generate wiki
files with tables of the reaction numbers in TUV and MCM/GECKO-A from a
template.
New function write_wiki to write the wiki files in setrxns.jl.
Remove duplicate of 4-octanone from branched ketones list in
MCM-GECKO-A.db. New teplate files for wiki pages for the MCMv3.x and
MCM/GECKO-A in src/data.